requirements.txt
fastapi
uvicorn[standard]
redis
faiss-cpu
prometheus-client 
numpy
pandas
joblib
torch
locust
scikit-learn